refactor: courseInfoTable컴포넌트 리팩토링
